Keep Logging in the ActiveModelSerializers namespace
