+++ docs/general/logging.md
@@ -1,6 +1,6 @@
 # Logging
 
-If we are using AMS on Rails app by default the `Rails.logger` will be used.
+If we are using ActiveModel::Serializers on Rails app by default the `Rails.logger` will be used.
 
 On a non Rails enviroment by default the `ActiveSupport::TaggedLogging` will be
 used.
@@ -8,5 +8,5 @@ used.
 If we need to customize the logger we can define this in an initializer:
 
 ```ruby
-ActiveModel::Serializer.logger = Logger.new(STDOUT)
+ActiveModelSerializers.logger = Logger.new(STDOUT)
 ```
