+++ lib/active_model/serializer.rb
@@ -6,7 +6,6 @@ require 'active_model/serializer/associations'
 require 'active_model/serializer/configuration'
 require 'active_model/serializer/fieldset'
 require 'active_model/serializer/lint'
-require 'active_model/serializer/logging'
 
 # ActiveModel::Serializer is an abstract class that is
 # reified when subclassed to decorate a resource.
@@ -14,7 +13,6 @@ module ActiveModel
   class Serializer
     include Configuration
     include Associations
-    include Logging
     require 'active_model/serializer/adapter'
 
     # Matches
