+++ test/active_model_serializers/logging_test.rb
